Module rustc_apfloat::ieee

Modules

sig 🔒

Implementation details of IeeeFloat significands, such as big integer arithmetic. As a rule of thumb, no functions in this module should dynamically allocate.

Macros

ieee_semantics 🔒

Structs

DoubleS

HalfS

IeeeFloat

QuadS

SingleS

X87DoubleExtendedS

Enums

Loss 🔒

Enum that represents what fraction of the LSB truncated bits of an fp number represent.

Constants

LIMB_BITS 🔒

Traits

Semantics

Represents floating point arithmetic semantics.

Functions

limbs_for_bits 🔒

Type Definitions

Double

Half

Limb 🔒

Fundamental unit of big integer arithmetic, but also large to store the largest significands by itself.

Quad

Single

X87DoubleExtended