Function rustc_apfloat::ieee::sig::each_chunk

pub(super) fn each_chunk<F: FnMut(u128) -> u128>(
    limbs: &mut [u128],
    bits: usize,
    f: F
)

For every consecutive chunk of bits bits from limbs, going from most significant to the least significant bits, call f to transform those bits and store the result back.