Trait rustc_apfloat::FloatConvert

pub trait FloatConvert<T: Float>: Float {
    fn convert_r(self, round: Round, loses_info: &mut bool) -> StatusAnd<T>;

    fn convert(self, loses_info: &mut bool) -> StatusAnd<T> { ... }
}

Required Methods

fn convert_r(self, round: Round, loses_info: &mut bool) -> StatusAnd<T>

Converts a value of one floating point type to another. The return value corresponds to the IEEE754 exceptions. *loses_info records whether the transformation lost information, i.e., whether converting the result back to the original type will produce the original value (this is almost the same as return value == Status::OK, but there are edge cases where this is not so).

Provided Methods

fn convert(self, loses_info: &mut bool) -> StatusAnd<T>