Module rustc_apfloat::ieee

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Structs

DoubleS 	 [
Experimental
]

HalfS 	 [
Experimental
]

IeeeFloat 	 [
Experimental
]

QuadS 	 [
Experimental
]

SingleS 	 [
Experimental
]

X87DoubleExtendedS 	 [
Experimental
]

Traits

Semantics 	 [
Experimental
]
Represents floating point arithmetic semantics.

Type Definitions

Double 	 [
Experimental
]

Half 	 [
Experimental
]

Quad 	 [
Experimental
]

Single 	 [
Experimental
]

X87DoubleExtended 	 [
Experimental
]