Struct rustc_apfloat::ieee::IeeeFloat

pub struct IeeeFloat<S> {
    sig: [u128; 1],
    exp: ExpInt,
    category: Category,
    sign: bool,
    marker: PhantomData<S>,
}

Fields
sig: [u128; 1]

Absolute significand value (including the integer bit).

exp: ExpInt

The signed unbiased exponent of the value.

category: Category

What kind of floating point number this is.

sign: bool

Sign bit of the number.

marker: PhantomData<S>
Implementations

impl<S: Semantics> IeeeFloat<S>

fn overflow_result(round: Round) -> StatusAnd<Self>

Handle positive overflow. We either return infinity or the largest finite number. For negative overflow, negate the round argument before calling.

fn round_away_from_zero(&self, round: Round, loss: Loss, bit: usize) -> bool

Returns true if, when truncating the current number, with bit the new LSB, with the given lost fraction and rounding mode, the result would need to be rounded away from zero (i.e., by increasing the signficand). This routine must work for Category::Zero of both signs, and Category::Normal numbers.

fn normalize(self, round: Round, loss: Loss) -> StatusAnd<Self>

fn from_hexadecimal_string(
s: &str,
round: Round
) -> Result<StatusAnd<Self>, ParseError>

fn from_decimal_string(
s: &str,
round: Round
) -> Result<StatusAnd<Self>, ParseError>