Trait rustc_apfloat::FloatConvert

pub trait FloatConvert<T: Float>: Float {
    fn convert_r(self, round: Round, loses_info: &mut bool) -> StatusAnd<T>;

    fn convert(self, loses_info: &mut bool) -> StatusAnd<T> { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn convert_r(self, round: Round, loses_info: &mut bool) -> StatusAnd<T>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Convert a value of one floating point type to another. The return value corresponds to the IEEE754 exceptions. *loses_info records whether the transformation lost information, i.e. whether converting the result back to the original type will produce the original value (this is almost the same as return value==Status::OK, but there are edge cases where this is not so).

Provided Methods

fn convert(self, loses_info: &mut bool) -> StatusAnd<T>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?